+++ src/media/mod.rs
@@ -8,6 +8,7 @@ pub mod frame_export;
 pub mod image;
 pub mod image_transform;
 pub mod metadata;
+pub mod metadata_writer;
 pub mod navigator;
 pub mod video;